requirements.txt
google-generativeai==0.7.2
sentence-transformers==2.7.0
faiss-cpu==1.8.0
pypdf==4.2.0
langchain-text-splitters==0.2.1
numpy==1.26.4
torch==2.3.1
transformers==4.41.2
python-dotenv==1.0.1
streamlit==1.35.0
python==3.11.8